Moving internal types to internal package
