+++ lib/observability/src/main/java/no/nordicsemi/memfault/observability/internal/db/ChunksDatabase.kt
@@ -29,28 +29,13 @@
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package no.nordicsemi.memfault.observability.db
+package no.nordicsemi.memfault.observability.internal.db
 
-import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.Query
-import kotlinx.coroutines.flow.Flow
+import androidx.room.Database
+import androidx.room.RoomDatabase
 
-@Dao
-internal interface ChunksDao {
+@Database(entities = [ChunkEntity::class], version = 1)
+internal abstract class ChunksDatabase : RoomDatabase() {
 
-    @Query("SELECT * FROM chunks WHERE device_id = :deviceId ORDER BY id DESC")
-    fun getAll(deviceId: String): Flow<List<ChunkEntity>>
-
-    @Query("SELECT * FROM chunks WHERE is_uploaded = 0 AND device_id = :deviceId ORDER BY id ASC LIMIT :limit")
-    fun getNotUploaded(limit: Int, deviceId: String): List<ChunkEntity>
-
-    @Query("UPDATE chunks SET is_uploaded = 1 WHERE is_uploaded IN (SELECT is_uploaded FROM chunks WHERE is_uploaded = 0 AND device_id = :deviceId ORDER BY id ASC LIMIT :limit)")
-    fun markUploaded(limit: Int, deviceId: String)
-
-    @Insert
-    fun insert(chunk: ChunkEntity)
-
-    @Query("DELETE FROM chunks WHERE is_uploaded = 1")
-    fun clearUploaded()
+    abstract fun chunksDao(): ChunksDao
 }
